Create Password generator
Description
This Python-based password generator allows users to create strong and random passwords. Users can specify the desired length of the password, and the generator will create a password using a combination of uppercase letters, lowercase letters, digits, and special characters.
How to Use
Run the Program: Execute the Python script in a Python environment.
Enter Password Length: The user is prompted to enter the desired length of the password.
View Generated Password: The program generates a random password of the specified length and displays it on the screen.
Repeat or Exit: The user can choose to generate another password or exit the program.
Features
Custom Length: Users can specify the length of the password.
Random Characters: The password is generated using a combination of uppercase letters, lowercase letters, digits, and special characters to ensure complexity and strength.
Error Handling: The program handles invalid input for password length and prompts the user to enter.